requirements.txt
anthropic>=0.49.0
openai>=1.58.0
mcp>=1.3.0
python-dotenv>=1.0.0
